# api/openapi.yaml
openapi: 3.1.0
info:
  title: Somra API
  version: 0.1.0-dev
  description: |
    Somra — all-in-one, self-hosted home media server.

    This document is the **design-first** source of truth for the Somra HTTP API.
    Frontend TypeScript types are generated from this file (see
    `scripts/gen-openapi-types.sh` and `make openapi-types`). The Go backend
    implements the contract defined here; any divergence is a bug in the
    implementation, not in this specification.

    Errors follow a `ProblemDetails`-style envelope: every error response uses
    the `ErrorResponse` schema, which is a superset compatible with RFC 7807
    "Problem Details for HTTP APIs" (the `messageKey` field carries an
    i18n key resolvable by the client; `message` carries the
    already-localized human text). See `plan/i18n-localization.md` for the
    binding i18n rules.
  license:
    name: AGPL-3.0-or-later
    identifier: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
  contact:
    name: Somra
    url: https://github.com/somralab/somra-media
servers:
  - url: /api/v1
    description: Default API root (relative to the Somra server host)
security:
  - {}
...
paths:
...
  /setup/admin:
    post:
      tags: [auth]
      summary: Create first admin (one-time)
      operationId: setupAdmin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupAdminRequest"
      responses:
        "201":
          description: Admin created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
...
  /auth/login:
    post:
      tags: [auth]
      summary: Login
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
...
  /libraries:
    get:
      tags: [library]
      summary: List libraries
      operationId: listLibraries
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Libraries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Library"
    post:
      tags: [library]
      summary: Create library
      operationId: createLibrary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LibraryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Library"
  /libraries/{libraryId}:
    parameters:
      - $ref: "#/components/parameters/libraryId"
    get:
      tags: [library]
      summary: Get library
      operationId: getLibrary
      responses:
        "200":
          description: Library
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Library"
    put:
      tags: [library]
      summary: Update library
      operationId: updateLibrary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LibraryInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Library"
    delete:
      tags: [library]
      summary: Delete library
      operationId: deleteLibrary
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Short-lived JWT access token issued by the auth module. The matching
        revocable refresh token is delivered out-of-band (HTTP-only cookie).
        Not applied to any operation in this revision; reserved for the auth
        endpoints introduced in Sprint 03.
  schemas:
...
    Library:
      type: object
      required: [id, name, kind, paths, watchEnabled]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        kind:
          type: string
          enum: [movie, tv, music]
        paths:
          type: array
          items:
            type: string
        watchEnabled:
          type: boolean
    LibraryInput:
      type: object
      required: [name, kind, paths]
      properties:
        name:
          type: string
        kind:
          type: string
          enum: [movie, tv, music]
        paths:
          type: array
          items:
            type: string
        watchEnabled:
          type: boolean
...
    SetupAdminRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
        deviceLabel:
          type: string
    TokenResponse:
      type: object
      required: [accessToken, expiresAt, user]
      properties:
        accessToken:
          type: string
        expiresAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
...
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        roles:
          type: array
          items:
            type: string
        disabled:
          type: boolean
...
  parameters:
    libraryId:
      name: libraryId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...